Create all ops before launching the graph
Love the examples!  

A small comment:

tf.initialize_all_variables() adds ops to  the graph to initialize the variables, as mentioned here 
http://tensorflow.org/how_tos/variables/index.html#initialization

In general, it's better practice to add all ops before launching the graph, as this avoids calling both DirectSession::Create() and then DirectSession::Extend() to add the new ops after the session is created.

Ref here:
https://github.com/tensorflow/tensorflow/blob/master/tensorflow/core/common_runtime/direct_session.cc#L161

It's not a huge deal, but might encourage others to do is that way too.